Up and Runing Environment. Still needs some glitch fixes. Metal Detector and UltrSonic AUtomation motion is ON.

+++ UltraSonic_driver/UltraSonic_priv.h
@@ -17,7 +17,7 @@
 typedef struct {
 	void (*fptr)(void);
 	Extint_enuIntNum Int_num;
-	u8 Val;
+	u16 Val;
 	UltraSonic_DeviceNumTriggerPin Trigger;
 }tstrdevicecreator;
